requirements.txt
streamlit
pandas
requests
python-dotenv
snowflake-connector-python
